Implement simple Scalar interpolations
Only parses and highlights as interpolation inside double-quoted
strings, as it should. No method call or indexer interpolation yet
(that will be a bit tricky due to the lookahead involved, and also
blocks on precedence limit handling).

This commit also adds some variables that will be used to control
over kinds of escape-y things in string literals.