Handle $.foo(...) style method calls
Add $*IN_DECL handling as part of this. This nicely means that in a
declaration context we can highlight $.foo like a variable, but in a
usage context we can highlight it as the method call that it really
is (highlighting the $ like self, the . as the method call ., and
the name as a method call name). This will make it stand out very
notably from $!foo, which is just a variable access.