Parse/highlight ** regex quantifier
Including handling of the codeblock case. Some discussion as to how we
highlight the ranges on a quantifier could be good; not sure whether to
highlight ** 1..^5 with the 1 and 5 as the integer color and ..^ as the
infix color, or 1..* with * as the whatever star color, or whether to
just have them all in the regex quantifier color. Of course, the block
case highlights those inside of the block in the way I described, since
then they are just normal code.