Implement traits, plus robustify grammar
This adds all the traits. While trying them out, I ran into a few bugs
where the parser got itself into a twist, so the remaining tweaks to
the grammar are to straighten a bunch of those out (separate lexing and
parsing can go wrong when the token sequences end up ambiguous).