Harden against backtracking due to <.ws> failure
IntelliJ can insert a fake token into the text it asks us to parse,
which is an identifier. Thus when the user types `if 5`, we can be
given `if 5IntelliJRulezzz`. Since <.ws> won't match between 5 and I
(due to <!ww> failing), we could backtrack in various places that were
making the bad assumption that <.ws> would never fail. Note that this
could also happen by a user making a typo or just editing stuff; of
note, typoing 5i (complex constant) as 5j (syntax error) could have
also ran into this.

Thus, I've audited the grammar and fixed up all the places where I
believe <.ws> might fail to allow for that.