Parse/highlight operator adverbs
These are a bit tricky, since they are parsed as fake infixes. In an
ideal world we'd end up with them incorporated into the node of the
operator they apply to, but that's rather tricky to arrange since they
show up "after the fact" (we'd need to notably complicate the reduce
process by introducing a "lag"). So, instead, have a node that carries
the adverb and what it's being applied to. This is good enough for us
to be able to do inspections for mistakes like `!%h<a>:exists`, where
it clearly represents in the tree that the adverb attaches to the !
prefix rather than the postcircumfix. By contrast, `not %h<a>:exists`
has it show up in the tree applied to the postcircumfix.