Refactor handling of scopes and declarations
This includes a far more generic and simple way of finding the
declarations made within a scope, which gets rid of various bits of
special-case code that would become a pain to maintain with time. This
also gives a clearer idea of scopes, which will be helpful to us in
adding support for Go To Declaration and similar. Finally, it means
that things like subs in subs now show up in structure view also, which
is a nice little improvement.