add annotation and quickfix for "unit" usage
complains about unit + braces and, more importantly,
about semicolon form without unit keyword.

It offers a quickfix that inserts the "unit" keyword,
but it doesn't seem to actually do anything if the file
in question is a scratch file.

Also, only offers the quickfix if the cursor is in
front of the semicolon. Should ideally offer whatever
place the cursor is on inside the declaration line
or something.

Part of P6I-147