Try to get the lexer to work incrementally
This does seem to somewhat help; now in a larger file we only lag when
we hit enter, which really does need a proper tree. Also, that should
be a bit faster than it was too, since we keep around a bunch less
resume state. This needs some real-world testing, so in a branch for
the time being.