Group comment tokens into a single PSI node
Most immediately so that the statement formatter sees it as a single
thing and not two things, and so won't insert a linebreak between the
`#` and the comment text. Granted we could have fixed this in some
different ways, but this feels like it gets us a cleaner tree overall.