Don't hand back tokens before startOffset
Even though we need to cover these again internally to be safe, we
should not hand them back. The highlighter engine does not check the
positions of tokens, and just blindly inserts them into its internal
representation. This meant that over time, we could build up dozens
of copies of the same token. This would lead to bogus inbalances of
open/close braces, which is why we started seeing weird behavior with
brace insertion.