Basic display of covered lines in source
Just shows it as a marker in the files that are open at the time of the
profile. Only marks the opening line of multi-line statements, so a bit
more work will be needed on that front later. Doesn't yet update it to
reflect when we open other files; that will need wiring up too.