Fix end of block continuation line handling
This is handled fine when we explicitly format code, but when we are
coding "live" and we are right at the end of the block, then the current
child is deemed to be the blockoid. Thus we'd end up not doing the
continuation indentation level.

This isn't a beautiful solution (read: hack), but it will probably be
right 99% of the time, and when it's not the problem will be a lack of
continuation indent where we'd want it, rather than incorrect use of
continuation indent. So it's an improvement on today, when it's always
wrong in that position, and shouldn't make anything that was right
before become wrong.