Prepare debugger for stepping support
Unfortunately, there's not yet support from them in the remote debug
library. This does, however, get us to the point where we forward the
request to a method on the debug thread object, which in the meantime
just does a resume. That's wrong, of course, though beats exploding with
an abstract method exception.