Don't lose newly added SDKs in Project Structure
The new project wizard would apply the newly added SDK, meaning that it
was written into the application level SDK table. The Project Structure
dialog would not do that when a new SDK was added, meaning that we never
put it into the application level table, meaning it was never considered
valid.