Big refactor of symbol resolution
Break it into lexical and MOP-based resolution, much like a Perl 6
compiler does things. Get a lot of simplification as a result, and
probably a performance win too, since we can avoid looking for a load of
things that are never going to be relevant.

There's still quite a bit to clean up; for one, the externals need to
learn to properly contribute symbols under the new scheme, and for
another I temporarily broke the handling of `trusts` (though I will
probably put it back in a rather different way than it was done
before).