Make SdkChangeListener common extension
Before only standalone Comma had it and we did not encounter issues,
because our Sdk symbol cache was not tied to a project. With external
symbols refactoring that uses project for external elements,
it might cause issues of invalidated project of an element from
cache was detected, which led to an exception. Making sure we always
invalidate cache on project change means we avoid the exception
and potentially able to not mix up symbols from different
versions of modules used from different projects etc.

t