Do not create a dangling reference to a disposed project
As symbol loading is asynchronous it is not impossible to
finish getting symbols while the project was already diposed,
for example, a test has ended. To avoid creation of a dangling
reference, check if the project was already disposed. It still
has a possibility of race with a project being disposed after the check,
but much rarer.