Build signed .dmg as part of the build process
This is a hack and needs a lot of care and love.
It has SSH disabled because it was doing crazy things -- but will want it to be reenabled in the future.

sshExec now takes an array instead of a string because it's the only sane thing to do

Also the comma_community_build.gant needs to do it as well