Make Perl6FileSystem follow rules of ArchiveFileSystem
Technically, it is not really an archive, but we are making
the code to believe so, using a dummy LightVirtualFile
as an "archive file". Next we resolve and provide files using
a script, which for now proves to be rather slow.
On the other hand, it is a step in the right direction,
as now we are able to get file listings in project view,
as well as show file sources.
Currently the state of things is still pretty buggy and
needs more care.