Accept `//` file paths as well as `/`
I am not entirely sure why that happens,
but when we are dealing with loading of a file
on repeated project opening (so the IDE
workspace setup is loaded from the saved state
of opened files, etc), URLs to the file
have two slashes at the start,
but when we are dealing with a fresh project
and opening external library files using other ways,
it only has a single slash. Thus the regex
accepting both options makes the first one
work.