Be smarter on detecting resolving methods on self
We did a simple check if the caller was `self` before,
but as Raku can use type objects as first-class values,
where this simple heuristic gives us false negative.
Here we use a more complex solution actually trying to resolve
to the package on which the method is called and compare
innermost package with it.