Make inferType @NotNull and avoid NPE
NameSuggester was assuming result of `inferType` cannot be null,
which resulted in a NPE for a simple subroutine case without explicit type.
As type inference can be used for a wide range of things, we don't want our
users to guess or spend time on checks. Thus add NotNull annotation
and prevent returning null in subroutine case, thus fixing an exception.