Resolve stub classes to implementation (if any)
The solution is simple: do not treat stubbed packages as "found" ones,
thus causing applicator to continue lookup.

An annotation complaining about packages/subs not stubbed would be welcome,
but currently our resolution is less than awesome and misses some "simple" cases, for example:

sub a { class A {} }
A; # <- and here Comma does not see anything

So defer it.