Use resolved type in method resolution
Which means we are able to resolve methods on types that are not in
scope (or in scope under the same name) at the point of the call.
(The test case here is a bit artificial, but the common case is of a
nested class in another module, which is given my its short name in a
return constraint. Previously, we would not give method completion and
resolution for it.)