Refactor Pod rendering into 2 stages
The first stage builds a Pod DOM that is a somewhat abstracted from the
details of syntactic processing. For example, all the dealings with
figuring out paragraph boundaries are gone at that level. This is then
turn into HTML in a second pass.

This hopefully has fixed a copule of minor bugs without introducing any
new ones. It certainly opens us up to writing tests for the PSI => Pod
DOM translation phase (I'll add some), and should make it possible to
implement a few more things (like item lists and tables) that are
currently missing without the code being utterly impossible to follow.