Fix heredoc's in an infix infinite re-indenting
If a heredoc is used withing an infix op,
before we wrongly set an alignment to it which
created additional indent, which was then re-parsed as
part of the heredoc and so needed even more indent ad nausea.
To avoid this just detect the situation and do not
set up an alignment if there is an infix present.