Use stub PSI when finding package traits
The child class line marker iterates the index of all types to see if
they have a `does` or `is` trait for the type in question. Despite us
having traits availble in the stub PSI, `findTrait` would look at the
real PSI only. That caused us to try parsing every file in the project
in order to calculate the line markers. Instead, for package decls, use
the stub PSI if it is available.