Provide cancellation of grammar preview
If one writes a grammar that won't terminate, then there was previously
no way to cancel it other than to kill the Rakudo process, or maybe to
look at a different grammar to make it "lose interest". Add a cancel
link while running the grammar to provide a better option.