Avoid eating up test output starting with a #
Prior to that output starting with '#' was parsed
by the TAP parser as a comment and we did not included them
in the resulting output.

This is both right and wrong, as first we don't want to display
"service" comments like a name of a subtest, but we also lose the
users output, so pick the lesser evil given TAP doesn't allow
us to differentiate.